Refactor package structure, remove unnecessary commented codes, simplify IoC Factory classes
