+++ BasicHyTools/src/main/java/hu/martinmarkus/basichytools/Main.java
@@ -1,6 +1,6 @@
 package hu.martinmarkus.basichytools;
 
-import hu.martinmarkus.basichytools.configmanagement.initializers.ModuleInitializer;
+import hu.martinmarkus.basichytools.initializers.ModuleInitializer;
 import hu.martinmarkus.basichytools.eventmanagement.ChatEventHandler;
 import hu.martinmarkus.basichytools.eventmanagement.CommandEventHandler;
 import hu.martinmarkus.basichytools.eventmanagement.UserConnectionEventHandler;
