--- BasicHyTools/src/main/java/hu/martinmarkus/basichytools/configmanagement/initializers/ioc/DefaultConfigFactory.java
@@ -1,9 +1,0 @@
-package hu.martinmarkus.basichytools.configmanagement.initializers.ioc;
-
-import hu.martinmarkus.basichytools.models.DefaultConfig;
-
-public class DefaultConfigFactory extends ObjectFactory<DefaultConfig> implements IDefaultConfigFactory {
-    public DefaultConfigFactory() {
-        super("defaultConfigSettings.xml");
-    }
-}
