--- BasicHyTools/src/main/java/hu/martinmarkus/basichytools/configmanagement/initializers/ioc/FunctionParameterFactory.java
@@ -1,9 +1,0 @@
-package hu.martinmarkus.basichytools.configmanagement.initializers.ioc;
-
-import hu.martinmarkus.basichytools.models.FunctionParameter;
-
-public class FunctionParameterFactory extends ObjectFactory<FunctionParameter> implements IFunctionParameterFactory {
-    public FunctionParameterFactory() {
-        super("functionParameterSettings.xml");
-    }
-}
