--- BasicHyTools/src/main/java/hu/martinmarkus/basichytools/configmanagement/initializers/ioc/IDefaultConfigFactory.java
@@ -1,6 +1,0 @@
-package hu.martinmarkus.basichytools.configmanagement.initializers.ioc;
-
-import hu.martinmarkus.basichytools.models.DefaultConfig;
-
-public interface IDefaultConfigFactory extends IObjectFactory<DefaultConfig> {
-}
