--- BasicHyTools/src/main/java/hu/martinmarkus/basichytools/configmanagement/initializers/ioc/IFunctionParameterFactory.java
@@ -1,6 +1,0 @@
-package hu.martinmarkus.basichytools.configmanagement.initializers.ioc;
-
-import hu.martinmarkus.basichytools.models.FunctionParameter;
-
-public interface IFunctionParameterFactory extends IObjectFactory<FunctionParameter> {
-}
