--- BasicHyTools/src/main/java/hu/martinmarkus/basichytools/configmanagement/initializers/ioc/IGameFunctionFactory.java
@@ -1,6 +1,0 @@
-package hu.martinmarkus.basichytools.configmanagement.initializers.ioc;
-
-import hu.martinmarkus.basichytools.gamefunctions.GameFunction;
-
-public interface IGameFunctionFactory extends IObjectFactory<GameFunction> {
-}
