--- BasicHyTools/src/main/java/hu/martinmarkus/basichytools/configmanagement/initializers/ioc/IGroupFactory.java
@@ -1,6 +1,0 @@
-package hu.martinmarkus.basichytools.configmanagement.initializers.ioc;
-
-import hu.martinmarkus.basichytools.models.Group;
-
-public interface IGroupFactory extends IObjectFactory<Group> {
-}
