--- BasicHyTools/src/main/java/hu/martinmarkus/basichytools/configmanagement/initializers/ioc/ILanguageConfigFactory.java
@@ -1,6 +1,0 @@
-package hu.martinmarkus.basichytools.configmanagement.initializers.ioc;
-
-import hu.martinmarkus.basichytools.models.LanguageConfig;
-
-public interface ILanguageConfigFactory extends IObjectFactory<LanguageConfig> {
-}
