--- BasicHyTools/src/main/java/hu/martinmarkus/basichytools/configmanagement/initializers/ioc/LanguageConfigFactory.java
@@ -1,9 +1,0 @@
-package hu.martinmarkus.basichytools.configmanagement.initializers.ioc;
-
-import hu.martinmarkus.basichytools.models.LanguageConfig;
-
-public class LanguageConfigFactory extends ObjectFactory<LanguageConfig> implements  ILanguageConfigFactory {
-    public LanguageConfigFactory() {
-        super("languageConfigSettings.xml");
-    }
-}
