+++ BasicHyTools/src/main/java/hu/martinmarkus/basichytools/eventmanagement/CommandEventHandler.java
@@ -1,17 +1,15 @@
 package hu.martinmarkus.basichytools.eventmanagement;
 
-import hu.martinmarkus.basichytools.configmanagement.initializers.ioc.GameFunctionFactory;
-import hu.martinmarkus.basichytools.configmanagement.initializers.ioc.IGameFunctionFactory;
-import hu.martinmarkus.basichytools.configmanagement.managers.DefaultConfigManager;
-import hu.martinmarkus.basichytools.configmanagement.managers.FunctionParameterManager;
-import hu.martinmarkus.basichytools.configmanagement.managers.LanguageConfigManager;
-import hu.martinmarkus.basichytools.configmanagement.managers.UserManager;
+import hu.martinmarkus.basichytools.initializers.iocfactories.concretefactories.GameFunctionFactory;
+import hu.martinmarkus.basichytools.configmanagement.DefaultConfigManager;
+import hu.martinmarkus.basichytools.configmanagement.FunctionParameterManager;
+import hu.martinmarkus.basichytools.configmanagement.LanguageConfigManager;
+import hu.martinmarkus.basichytools.configmanagement.UserManager;
 import hu.martinmarkus.basichytools.gamefunctions.GameFunction;
 import hu.martinmarkus.basichytools.models.DefaultConfig;
 import hu.martinmarkus.basichytools.models.FunctionParameter;
 import hu.martinmarkus.basichytools.models.LanguageConfig;
 import hu.martinmarkus.basichytools.models.User;
-import hu.martinmarkus.basichytools.configmanagement.initializers.ioc.ObjectFactory;
 
 import java.util.List;
 
@@ -79,7 +77,7 @@ public class CommandEventHandler {
     private GameFunction defineGameFunction(String rawCommand) {
         String command = getCommand(rawCommand);
         List<FunctionParameter> functionParameters = getFunctionParameters();
-        IGameFunctionFactory gameFunctionFactory = GameFunctionFactory.getInstance();
+        GameFunctionFactory gameFunctionFactory = GameFunctionFactory.getInstance();
 
         for (FunctionParameter functionParameter : functionParameters) {
             String name = functionParameter.getName().toLowerCase();
