+++ BasicHyTools/src/main/java/hu/martinmarkus/basichytools/globalmechanisms/chatmechanisms/Announcer.java
@@ -1,6 +1,6 @@
 package hu.martinmarkus.basichytools.globalmechanisms.chatmechanisms;
 
-import hu.martinmarkus.basichytools.configmanagement.managers.DefaultConfigManager;
+import hu.martinmarkus.basichytools.configmanagement.DefaultConfigManager;
 import hu.martinmarkus.basichytools.models.DefaultConfig;
 
 import java.util.List;
