+++ BasicHyTools/src/main/java/hu/martinmarkus/basichytools/globalmechanisms/chatmechanisms/Informer.java
@@ -1,6 +1,6 @@
 package hu.martinmarkus.basichytools.globalmechanisms.chatmechanisms;
 
-import hu.martinmarkus.basichytools.configmanagement.managers.LanguageConfigManager;
+import hu.martinmarkus.basichytools.configmanagement.LanguageConfigManager;
 import hu.martinmarkus.basichytools.utils.DateTimeUtil;
 
 public class Informer {
