+++ BasicHyTools/src/main/java/hu/martinmarkus/basichytools/initializers/iocfactories/concretefactories/DefaultConfigFactory.java
@@ -1,0 +1,10 @@
+package hu.martinmarkus.basichytools.initializers.iocfactories.concretefactories;
+
+import hu.martinmarkus.basichytools.initializers.iocfactories.ObjectFactory;
+import hu.martinmarkus.basichytools.models.DefaultConfig;
+
+public class DefaultConfigFactory extends ObjectFactory<DefaultConfig> {
+    public DefaultConfigFactory() {
+        super("defaultConfigSettings.xml");
+    }
+}
