+++ BasicHyTools/src/main/java/hu/martinmarkus/basichytools/initializers/iocfactories/concretefactories/GroupFactory.java
@@ -1,0 +1,10 @@
+package hu.martinmarkus.basichytools.initializers.iocfactories.concretefactories;
+
+import hu.martinmarkus.basichytools.initializers.iocfactories.ObjectFactory;
+import hu.martinmarkus.basichytools.models.Group;
+
+public class GroupFactory extends ObjectFactory<Group> {
+    public GroupFactory() {
+        super("groupsSettings.xml");
+    }
+}
