+++ BasicHyTools/src/main/java/hu/martinmarkus/basichytools/initializers/iocfactories/concretefactories/LanguageConfigFactory.java
@@ -1,0 +1,10 @@
+package hu.martinmarkus.basichytools.initializers.iocfactories.concretefactories;
+
+import hu.martinmarkus.basichytools.initializers.iocfactories.ObjectFactory;
+import hu.martinmarkus.basichytools.models.LanguageConfig;
+
+public class LanguageConfigFactory extends ObjectFactory<LanguageConfig> {
+    public LanguageConfigFactory() {
+        super("languageConfigSettings.xml");
+    }
+}
