+++ BasicHyTools/src/main/java/hu/martinmarkus/basichytools/models/User.java
@@ -1,12 +1,12 @@
 package hu.martinmarkus.basichytools.models;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import hu.martinmarkus.basichytools.configmanagement.managers.DefaultConfigManager;
-import hu.martinmarkus.basichytools.configmanagement.managers.GroupManager;
-import hu.martinmarkus.basichytools.configmanagement.managers.LanguageConfigManager;
+import hu.martinmarkus.basichytools.configmanagement.DefaultConfigManager;
+import hu.martinmarkus.basichytools.configmanagement.GroupManager;
+import hu.martinmarkus.basichytools.configmanagement.LanguageConfigManager;
 import hu.martinmarkus.basichytools.utils.PlaceholderReplacer;
-import hu.martinmarkus.basichytools.permissionmanagement.PermissionValidator;
-import hu.martinmarkus.basichytools.permissionmanagement.UserPermissionValidator;
+import hu.martinmarkus.basichytools.utils.permissionmanagement.PermissionValidator;
+import hu.martinmarkus.basichytools.utils.permissionmanagement.UserPermissionValidator;
 import hu.martinmarkus.basichytools.utils.StringUtil;
 
 import java.util.ArrayList;
