+++ BasicHyTools/src/main/java/hu/martinmarkus/basichytools/utils/permissionmanagement/PartCheckResult.java
@@ -1,4 +1,4 @@
-package hu.martinmarkus.basichytools.permissionmanagement;
+package hu.martinmarkus.basichytools.utils.permissionmanagement;
 
 class PartCheckResult {
     private boolean concretePermission;
